Made `register` bound to a specific Express instance.
Instead of changing the settings at the module level, we allow the user to create a new instance of `partials` for each Express instance.
This avoid issues with conflicting settings when multiple Express engines are running concurrently.

In that case the usage would then be:

    var express_partials = require('express-partials');

    /* express1 uses jade */
    express1.set('view engine','jade');
    partials1 = express_partials();
    partials1.register('html','jade');
    express1.use(partials1);

    /* express2 uses ejs */
    express2.set('view engine','ejs');
    partials2 = express_partials();
    partials2.register('html','ejs');
    express2.use(partials1);

Without this patch, the latest `register('html',...')` called would override the options for both.